fix(writer): preserve workflow capabilities after rebase
