--- algorithm/lazymind/chat/engine/prompts/writer_media.py
@@ -1,8 +1,0 @@
-WRITER_IMAGE_ACQUISITION_PROMPT = '''Create one professional visual for a long-form document.
-
-Visual type: {visual_type}
-The visual must communicate: {purpose}
-
-Keep the composition clear and suitable for insertion into a document. Avoid watermarks,
-brand logos, decorative filler, and small unreadable text. Return exactly one image.
-'''
