+++ frontend/src/i18n/locales/en-US.ts
@@ -1756,10 +1756,11 @@ const enUS = {
       writingDetails: "Writing details",
       writingLanguage: "Writing language",
     },
-    pluginStatusRunning: "Running",
-    pluginStatusDone: "Done",
-    pluginStatusWaiting: "Waiting for approval",
-    pluginStatusFailed: "Failed",
+    workflowStatusRunning: "Running",
+    workflowStatusDone: "Done",
+    workflowStatusWaiting: "Waiting for approval",
+    workflowStatusFailed: "Failed",
+    workflowStatusStopped: "Stopped",
     pluginBtnDisabledHint: "Step is running, please wait…",
     pluginSavingBeforeAction: "Saving edits…",
     pluginRetryFlushHint: "Saves current edits, then retries",
