+++ frontend/src/i18n/locales/zh-CN.ts
@@ -1584,10 +1584,11 @@ const zhCN = {
     pluginRowAria: "第 {{index}} 行",
     pluginImageAria: "图片 {{index}}",
     pluginContentItemAria: "内容项 {{index}}",
-    pluginStatusRunning: "执行中",
-    pluginStatusDone: "已完成",
-    pluginStatusWaiting: "等待审批",
-    pluginStatusFailed: "失败",
+    workflowStatusRunning: "执行中",
+    workflowStatusDone: "已完成",
+    workflowStatusWaiting: "等待审批",
+    workflowStatusFailed: "失败",
+    workflowStatusStopped: "已停止",
     pluginStatusAria: "工作流状态：{{status}}",
     pluginViewWorkflow: "查看工作流",
     pluginSessionControls: "工作流会话操作",
